move exception.hpp to common
