+++ power_grid_model_c/power_grid_model/include/power_grid_model/auxiliary/dataset_handler.hpp
@@ -8,8 +8,8 @@
 
 // handle dataset and buffer related stuff
 
+#include "../common/exception.hpp"
 #include "../common/power_grid_model.hpp"
-#include "../exception.hpp"
 #include "dataset.hpp"
 #include "meta_data.hpp"
 #include "meta_data_gen.hpp"
