+++ power_grid_model_c/power_grid_model/include/power_grid_model/auxiliary/meta_data.hpp
@@ -7,9 +7,9 @@
 #define POWER_GRID_MODEL_META_DATA_HPP
 
 #include "../common/enum.hpp"
+#include "../common/exception.hpp"
 #include "../common/power_grid_model.hpp"
 #include "../common/three_phase_tensor.hpp"
-#include "../exception.hpp"
 
 #include <bit>
 #include <span>
