+++ power_grid_model_c/power_grid_model/include/power_grid_model/common/grouped_index_vector.hpp
@@ -6,8 +6,8 @@
 #ifndef POWER_GRID_MODEL_SPARSE_IDX_VECTOR_HPP
 #define POWER_GRID_MODEL_SPARSE_IDX_VECTOR_HPP
 
-#include "common/power_grid_model.hpp"
-#include "common/typing.hpp"
+#include "power_grid_model.hpp"
+#include "typing.hpp"
 
 #include <boost/iterator/iterator_facade.hpp>
 #include <boost/iterator/zip_iterator.hpp>
