+++ power_grid_model_c/power_grid_model/include/power_grid_model/component/branch3.hpp
@@ -12,7 +12,7 @@
 #include "../auxiliary/output.hpp"
 #include "../auxiliary/update.hpp"
 #include "../calculation_parameters.hpp"
-#include "../exception.hpp"
+#include "../common/exception.hpp"
 
 namespace power_grid_model {
