+++ power_grid_model_c/power_grid_model/include/power_grid_model/component/fault.hpp
@@ -13,7 +13,7 @@
 #include "../auxiliary/update.hpp"
 #include "../calculation_parameters.hpp"
 #include "../common/enum.hpp"
-#include "../exception.hpp"
+#include "../common/exception.hpp"
 
 namespace power_grid_model {
