+++ power_grid_model_c/power_grid_model/include/power_grid_model/component/link.hpp
@@ -12,9 +12,9 @@
 #include "../auxiliary/output.hpp"
 #include "../auxiliary/update.hpp"
 #include "../calculation_parameters.hpp"
+#include "../common/exception.hpp"
 #include "../common/power_grid_model.hpp"
 #include "../common/three_phase_tensor.hpp"
-#include "../exception.hpp"
 
 namespace power_grid_model {
