+++ power_grid_model_c/power_grid_model/include/power_grid_model/component/power_sensor.hpp
@@ -10,8 +10,8 @@
 
 #include "../calculation_parameters.hpp"
 #include "../common/enum.hpp"
+#include "../common/exception.hpp"
 #include "../common/power_grid_model.hpp"
-#include "../exception.hpp"
 
 namespace power_grid_model {
