+++ power_grid_model_c/power_grid_model/include/power_grid_model/component/voltage_sensor.hpp
@@ -9,9 +9,9 @@
 #include "sensor.hpp"
 
 #include "../calculation_parameters.hpp"
+#include "../common/exception.hpp"
 #include "../common/power_grid_model.hpp"
 #include "../common/three_phase_tensor.hpp"
-#include "../exception.hpp"
 
 #include <limits>
