+++ power_grid_model_c/power_grid_model/include/power_grid_model/math_solver/common_solver_functions.hpp
@@ -10,7 +10,7 @@
 #include "y_bus.hpp"
 
 #include "../calculation_parameters.hpp"
-#include "../exception.hpp"
+#include "../common/exception.hpp"
 
 namespace power_grid_model::math_solver::detail {
