+++ power_grid_model_c/power_grid_model/include/power_grid_model/math_solver/iterative_current_pf_solver.hpp
@@ -61,9 +61,9 @@ Nomenclature:
 #include "y_bus.hpp"
 
 #include "../calculation_parameters.hpp"
+#include "../common/exception.hpp"
 #include "../common/power_grid_model.hpp"
 #include "../common/three_phase_tensor.hpp"
-#include "../exception.hpp"
 #include "../timer.hpp"
 
 namespace power_grid_model::math_solver {
