+++ power_grid_model_c/power_grid_model/include/power_grid_model/math_solver/linear_pf_solver.hpp
@@ -35,9 +35,9 @@ if there are sources
 #include "y_bus.hpp"
 
 #include "../calculation_parameters.hpp"
+#include "../common/exception.hpp"
 #include "../common/power_grid_model.hpp"
 #include "../common/three_phase_tensor.hpp"
-#include "../exception.hpp"
 #include "../timer.hpp"
 
 namespace power_grid_model::math_solver {
