+++ power_grid_model_c/power_grid_model/include/power_grid_model/math_solver/newton_raphson_pf_solver.hpp
@@ -152,9 +152,9 @@ J.L -= -dQ_cal_m/dV
 #include "y_bus.hpp"
 
 #include "../calculation_parameters.hpp"
+#include "../common/exception.hpp"
 #include "../common/power_grid_model.hpp"
 #include "../common/three_phase_tensor.hpp"
-#include "../exception.hpp"
 #include "../timer.hpp"
 
 namespace power_grid_model::math_solver {
