+++ tests/cpp_unit_tests/test_math_solver.cpp
@@ -4,8 +4,8 @@
 
 // In this unit test the powerflow, state estimation and short circuit solvers are tested
 
+#include <power_grid_model/common/exception.hpp>
 #include <power_grid_model/common/three_phase_tensor.hpp>
-#include <power_grid_model/exception.hpp>
 #include <power_grid_model/math_solver/math_solver.hpp>
 #include <power_grid_model/math_solver/newton_raphson_pf_solver.hpp>
